services/classification-agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
clip @ git+https://github.com/openai/CLIP.git
